+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 
-- Added reply shortcuts to Message object
+- Added reply shortcuts to message
+- Added answer shortcuts to callback, inline, pre-checkout and shipping queries
 
 ## 0.0.43